requirement.txt
1.bs4 
2.BeautifulSoup
3.selenium
4.webdriver